+++ Applications/DataExplorer/VtkVis/VtkRaster.cpp
@@ -180,9 +180,9 @@ vtkImageImport* VtkRaster::loadImageFromTIFF(const std::string &fileName,
                     if (photometric==1 && colormap_used==1)
                     {
                         int idx = TIFFGetR(pixVal[pxl_idx]);
-                        data[pos]   = cmap_red[idx] >> 8;
-                        data[pos+1] = cmap_green[idx] >> 8;
-                        data[pos+2] = cmap_blue[idx] >> 8;
+                        data[pos]   = static_cast<float>(cmap_red[idx] >> 8);
+                        data[pos+1] = static_cast<float>(cmap_green[idx] >> 8);
+                        data[pos+2] = static_cast<float>(cmap_blue[idx] >> 8);
                         data[pos+3] = 1;
                     }
                     else
